+++ stroom-config/stroom-config-global-impl-db/src/main/resources/stroom/config/global/impl/db/migration/V07_01_00_001__preferences.sql
@@ -1,0 +1,36 @@
+-- ------------------------------------------------------------------------
+-- Copyright 2020 Crown Copyright
+--
+-- Licensed under the Apache License, Version 2.0 (the "License");
+-- you may not use this file except in compliance with the License.
+-- You may obtain a copy of the License at
+--
+--     http://www.apache.org/licenses/LICENSE-2.0
+--
+-- Unless required by applicable law or agreed to in writing, software
+-- distributed under the License is distributed on an "AS IS" BASIS,
+-- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+-- See the License for the specific language governing permissions and
+-- limitations under the License.
+-- ------------------------------------------------------------------------
+
+-- Stop NOTE level warnings about objects (not)? existing
+SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0;
+
+--
+-- Create the config table
+--
+CREATE TABLE IF NOT EXISTS preferences (
+    id                    int NOT NULL AUTO_INCREMENT,
+    version               int NOT NULL,
+    create_time_ms        bigint NOT NULL,
+    create_user           varchar(255) NOT NULL,
+    update_time_ms        bigint NOT NULL,
+    update_user           varchar(255) NOT NULL,
+    user_id               varchar(255) NOT NULL,
+    dat                   longtext NOT NULL,
+    PRIMARY KEY           (id),
+    UNIQUE KEY            user_id (user_id)
+) ENGINE=InnoDB DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci;
+
+SET SQL_NOTES=@OLD_SQL_NOTES;