+++ stroom-config/stroom-config-global-impl/src/main/java/stroom/config/global/impl/PreferencesDao.java
@@ -1,0 +1,12 @@
+package stroom.config.global.impl;
+
+import stroom.ui.config.shared.UserPreferences;
+
+public interface PreferencesDao {
+
+    UserPreferences fetch(String userId);
+
+    int update(String userId, UserPreferences userPreferences);
+
+    int delete(String userId);
+}