+++ stroom-dashboard-gwt/src/main/java/stroom/dashboard/client/gin/DashboardAppGinjector.java
@@ -27,6 +27,8 @@ import stroom.dashboard.client.vis.gin.VisModule;
 import stroom.dispatch.client.RestModule;
 import stroom.entity.client.gin.EntityGinjector;
 import stroom.entity.client.gin.EntityModule;
+import stroom.preferences.client.gin.PreferencesGinjector;
+import stroom.preferences.client.gin.PreferencesModule;
 import stroom.query.client.QueryModule;
 import stroom.security.client.gin.SecurityGinjector;
 import stroom.security.client.gin.SecurityModule;
@@ -48,6 +50,7 @@ import com.gwtplatform.mvp.client.proxy.PlaceManager;
         AlertModule.class,
         SecurityModule.class,
         EntityModule.class,
+        PreferencesModule.class,
         QueryModule.class,
         DashboardModule.class,
         VisModule.class})
@@ -58,6 +61,7 @@ public interface DashboardAppGinjector extends
         SecurityGinjector,
         EntityGinjector,
         DashboardGinjector,
+        PreferencesGinjector,
         VisGinjector {
 
     // Default implementation of standard resources